Backend/requirements.txt
fastapi
uvicorn
pillow
pytesseract
scikit-learn
numpy
python-multipart